backend/requirements.txt
fastapi
uvicorn
python-dotenv
openai
requests
python-jose[cryptography]
google-auth
pydantic
python-multipart
pymongo
langchain
langchain-openai
langgraph
langchain-community
